requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
-e git+ssh://git@github.com/zbrooks442/gke-peering-event-function.git@52dec0de2ff13fd9cb58c6e923a016b729efc878#egg=gcp_gke_vpc_updater
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-compute==1.11.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
idna==3.4
mock==5.0.2
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.29.0
rsa==4.9
six==1.16.0
urllib3==1.26.15
